+++ example_with_integration_test/lib/models/todo_model.g.dart
@@ -10,7 +10,7 @@ Todo _$TodoFromJson(Map<String, dynamic> json) => Todo(
       id: json['id'] as String,
       created: DateTime.parse(json['created'] as String),
       updated: DateTime.parse(json['updated'] as String),
-      status: _$enumDecode(_$TodoStatusEnumMap, json['status']),
+      status: $enumDecode(_$TodoStatusEnumMap, json['status']),
       action: json['action'] as String?,
     );
 
@@ -21,32 +21,6 @@ Map<String, dynamic> _$TodoToJson(Todo instance) => <String, dynamic>{
       'action': instance.action,
       'status': _$TodoStatusEnumMap[instance.status],
     };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
 
 const _$TodoStatusEnumMap = {
   TodoStatus.pending: 'pending',