+++ lib/src/flutter/code_generation/generators/gherkin_suite_test_generator.dart
@@ -16,7 +16,7 @@ class GherkinSuiteTestGenerator
 class _CustomGherkinIntegrationTestRunner extends GherkinIntegrationTestRunner {
   _CustomGherkinIntegrationTestRunner(
     TestConfiguration configuration,
-    void Function(World) appMainFunction,
+    Future<void> Function(World) appMainFunction,
   ) : super(configuration, appMainFunction);
 
   @override
@@ -29,7 +29,7 @@ class _CustomGherkinIntegrationTestRunner extends GherkinIntegrationTestRunner {
 
 void executeTestSuite(
   TestConfiguration configuration,
-  void Function(World) appMainFunction,
+  Future<void> Function(World) appMainFunction,
 ) {
   _CustomGherkinIntegrationTestRunner(configuration, appMainFunction).run();
 }